+++ resources/release-notes.md
@@ -1,4 +1,5 @@
 ## Software
 
+- (All) Fix issue with loading events from `mag.TimeSeries` in `mag.graphics.chart.Stackedplot` (when `EventsVisible = true`)
 - (IMAP) Support old IMAP science file format (file time defined as `HH'h'` instead of `HH'h'mm`)
 - (IMAP) Fix typos is power health check descriptions